requirements.txt
tensorflow==2.0.0b0
Pillow==6.0.0
azure-storage-blob